requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.4
cssselect2==0.8.0
cymem==2.0.11
distlib==0.3.9
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-browser-reload==1.18.0
django-cors-headers==4.7.0
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.16.1
fonttools==4.58.4
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
numpy==2.3.0
packaging==25.0
pdfminer.six==20250506
pillow==11.2.1
platformdirs==4.3.6
preshed==3.0.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.9.0
PyMuPDF==1.26.1
pyphen==0.17.2
python-decouple==3.8
requests==2.32.4
rich==14.0.0
setuptools==80.9.0
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
thinc==8.3.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.28.1
wasabi==1.1.3
weasel==0.4.1
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
wrapt==1.17.2
zopfli==0.2.3.post1
